Move `git_streaming()` to `git_command`
